requirements.txt
torch
flask
request
argparse
pillow
os
io
